Fix: remove unused code
