+++ backend/WebUI/api_webui.go
@@ -22,9 +22,6 @@ import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"golang.org/x/crypto/bcrypt"
 
-	"github.com/free5gc/chf/cdr/asn"
-	"github.com/free5gc/chf/cdr/cdrFile"
-	"github.com/free5gc/chf/cdr/cdrType"
 	"github.com/free5gc/openapi/models"
 	"github.com/free5gc/util/mongoapi"
 	"github.com/free5gc/webconsole/backend/logger"
@@ -1930,283 +1927,3 @@ func OptionsSubscribers(c *gin.Context) {
 
 	c.JSON(http.StatusNoContent, gin.H{})
 }
-
-// Get vol from CDR
-// TS 32.297: Charging Data Record (CDR) file format and transfer
-func parseCDR(supi string) (map[int64]RatingGroupDataUsage, error) {
-	logger.BillingLog.Traceln("parseCDR")
-	fileName := "/tmp/webconsole/" + supi + ".cdr"
-	if _, err := os.Stat(fileName); err != nil {
-		return nil, err
-	}
-
-	newCdrFile := cdrFile.CDRFile{}
-
-	newCdrFile.Decoding(fileName)
-	dataUsage := make(map[int64]RatingGroupDataUsage)
-
-	for _, cdr := range newCdrFile.CdrList {
-		recvByte := cdr.CdrByte
-		val := reflect.New(reflect.TypeOf(&cdrType.ChargingRecord{}).Elem()).Interface()
-		err := asn.UnmarshalWithParams(recvByte, val, "")
-		if err != nil {
-			logger.BillingLog.Errorf("parseCDR error when unmarshal with params: %+v", err)
-			continue
-		}
-
-		chargingRecord := *(val.(*cdrType.ChargingRecord))
-
-		for _, multipleUnitUsage := range chargingRecord.ListOfMultipleUnitUsage {
-			rg := multipleUnitUsage.RatingGroup.Value
-			du := dataUsage[rg]
-
-			du.Snssai = fmt.Sprintf("%02d", chargingRecord.PDUSessionChargingInformation.NetworkSliceInstanceID.SST.Value) +
-				string(chargingRecord.PDUSessionChargingInformation.NetworkSliceInstanceID.SD.Value)
-
-			du.Dnn = string(chargingRecord.PDUSessionChargingInformation.DataNetworkNameIdentifier.Value)
-
-			for _, usedUnitContainer := range multipleUnitUsage.UsedUnitContainers {
-				du.TotalVol += usedUnitContainer.DataTotalVolume.Value
-				du.UlVol += usedUnitContainer.DataVolumeUplink.Value
-				du.DlVol += usedUnitContainer.DataVolumeDownlink.Value
-			}
-
-			dataUsage[rg] = du
-		}
-	}
-
-	return dataUsage, nil
-}
-
-func GetChargingRecord(c *gin.Context) {
-	logger.BillingLog.Info("Get Charging Record")
-	setCorsHeader(c)
-
-	if !CheckAuth(c) {
-		c.JSON(http.StatusUnauthorized, gin.H{"cause": "Illegal Token"})
-		return
-	}
-
-	webuiSelf := webui_context.GetSelf()
-	webuiSelf.UpdateNfProfiles()
-
-	// Get supi of UEs
-	var uesJsonData interface{}
-	if amfUris := webuiSelf.GetOamUris(models.NfType_AMF); amfUris != nil {
-		requestUri := fmt.Sprintf("%s/namf-oam/v1/registered-ue-context", amfUris[0])
-
-		res, err := http.NewRequestWithContext(context.TODO(), http.MethodGet, requestUri, nil)
-		if err != nil {
-			logger.ProcLog.Error(err)
-			c.JSON(http.StatusInternalServerError, gin.H{})
-			return
-		}
-		resp, res_err := httpsClient.Do(res)
-		if res_err != nil {
-			logger.ProcLog.Error(err)
-			c.JSON(http.StatusInternalServerError, gin.H{})
-			return
-		}
-
-		defer func() {
-			if closeErr := resp.Body.Close(); closeErr != nil {
-				logger.ProcLog.Error(closeErr)
-			}
-		}()
-
-		err = json.NewDecoder(resp.Body).Decode(&uesJsonData)
-		if err != nil {
-			logger.BillingLog.Error(err)
-		}
-	}
-
-	// build charging records
-	uesBsonA := toBsonA(uesJsonData)
-	chargingRecordsBsonA := make([]interface{}, 0, len(uesBsonA))
-
-	type OfflinePDUTypeMap struct {
-		supi           string
-		snssai         string
-		dnn            string
-		unitcost       int64
-		flowTotalVolum int64
-		flowTotalUsage int64
-	}
-
-	offlineChargingPDUTypeMap := make(map[string]OfflinePDUTypeMap)
-
-	for _, ueData := range uesBsonA {
-		ueBsonM := toBsonM(ueData)
-
-		supi := ueBsonM["Supi"].(string)
-
-		ratingGroupDataUsages, err := parseCDR(supi)
-		if err != nil {
-			logger.BillingLog.Warnln(err)
-			continue
-		}
-
-		for rg, du := range ratingGroupDataUsages {
-			filter := bson.M{"ratingGroup": rg}
-			chargingDataInterface, err := mongoapi.RestfulAPIGetOne(chargingDataColl, filter)
-			if err != nil {
-				logger.ProcLog.Errorf("PostSubscriberByID err: %+v", err)
-			}
-			if len(chargingDataInterface) == 0 {
-				logger.BillingLog.Warningf("ratingGroup: %d not found in mongoapi, may change the rg id", rg)
-				continue
-			}
-
-			var chargingData ChargingData
-			err = json.Unmarshal(mapToByte(chargingDataInterface), &chargingData)
-			if err != nil {
-				logger.BillingLog.Error(err)
-			}
-			logger.BillingLog.Debugf("add ratingGroup: %d, supi: %s, method: %s", rg, supi, chargingData.ChargingMethod)
-
-			switch chargingData.ChargingMethod {
-			case "Offline":
-				unitcost, err := strconv.ParseInt(chargingData.UnitCost, 10, 64)
-				if err != nil {
-					logger.BillingLog.Error("Offline unitCost strconv: ", err.Error())
-					unitcost = 1
-				}
-
-				// TODO(ctfang): add DNN back
-				key := chargingData.UeId + chargingData.Snssai
-				// key := chargingData.UeId + chargingData.Snssai + chargingData.Dnn
-				pdu_level, exist := offlineChargingPDUTypeMap[key]
-				if !exist {
-					pdu_level = OfflinePDUTypeMap{}
-					logger.BillingLog.Errorln("Create: ", key)
-				}
-				if chargingData.Filter != "" {
-					// Flow-based charging
-					du.Usage = du.TotalVol * unitcost
-
-					pdu_level.flowTotalUsage += du.Usage
-					pdu_level.flowTotalVolum += du.TotalVol
-
-					logger.BillingLog.Errorln(pdu_level.flowTotalUsage, pdu_level.flowTotalVolum)
-				} else {
-					// PDU-Session Based charging (DNN-Based)
-					pdu_level.snssai = chargingData.Snssai
-					pdu_level.dnn = chargingData.Dnn
-					pdu_level.supi = chargingData.UeId
-					pdu_level.unitcost = unitcost
-
-					logger.BillingLog.Errorln("PDU", pdu_level.flowTotalUsage, pdu_level.flowTotalVolum)
-				}
-				offlineChargingPDUTypeMap[key] = pdu_level
-			case "Online":
-				tmpInt, err1 := strconv.Atoi(chargingData.Quota)
-				if err1 != nil {
-					logger.BillingLog.Error("Quota strconv: ", err1, rg, du, chargingData)
-				}
-				du.QuotaLeft = int64(tmpInt)
-			}
-			du.Snssai = chargingData.Snssai
-			du.Dnn = chargingData.Dnn
-			du.Supi = supi
-			du.Filter = chargingData.Filter
-
-			ratingGroupDataUsages[rg] = du
-			chargingRecordsBsonA = append(chargingRecordsBsonA, toBsonM(du))
-		}
-
-		// logger.BillingLog.Debugln("addVolumeToPduLevelDataUsage")
-		// Flow's Tol, Dl, Ul volume add back to PDU Session Level
-		// TODO: O(n^2) too slow
-		// for _, du := range ratingGroupDataUsages {
-		// 	if du.Filter != "" {
-		// 		logger.BillingLog.Debugln(du)
-		// 		addVolumeToPduLevelDataUsage(ratingGroupDataUsages, du)
-		// 	}
-		// 	chargingRecordsBsonA = append(chargingRecordsBsonA, toBsonM(du))
-		// }
-	}
-	for idx, record := range chargingRecordsBsonA {
-		var rd RatingGroupDataUsage
-
-		tmp, err := json.Marshal(record)
-		if err != nil {
-			logger.BillingLog.Errorln("Marshal chargingRecordsBsonA error:", err.Error())
-			continue
-		}
-		err = json.Unmarshal(tmp, &rd)
-		if err != nil {
-			logger.BillingLog.Errorln("Unmarshall chargingRecordsBsonA error:", err.Error())
-			continue
-		}
-
-		if rd.Filter != "" {
-			continue
-		}
-
-		key := rd.Supi + rd.Snssai + rd.Dnn
-		if val, exist := offlineChargingPDUTypeMap[key]; exist {
-			rd.Usage += val.flowTotalUsage
-			rd.Usage += (rd.TotalVol - val.flowTotalVolum) * val.unitcost
-			chargingRecordsBsonA[idx] = toBsonM(rd)
-
-			logger.BillingLog.Warningln(val.flowTotalUsage, val.unitcost)
-			logger.BillingLog.Warningln(idx, rd.Usage)
-		}
-	}
-
-	c.JSON(http.StatusOK, chargingRecordsBsonA)
-}
-
-// nolint
-// func addVolumeToPduLevelDataUsage(ratingGroupDataUsages map[int64]RatingGroupDataUsage, flow_du RatingGroupDataUsage) {
-// 	for rg, du := range ratingGroupDataUsages {
-// 		if du.Filter == "" && du.Dnn == "" && du.Supi == flow_du.Supi && du.Snssai == flow_du.Snssai {
-// 			logger.BillingLog.Warningln("found usage: ", flow_du.Usage, du.Usage)
-// 			logger.BillingLog.Warningln("found volumn: ", flow_du.TotalVol, du.TotalVol)
-
-// 			du.TotalVol += flow_du.TotalVol
-// 			du.DlVol += flow_du.DlVol
-// 			du.UlVol += flow_du.UlVol
-// 			du.Usage += flow_du.Usage
-// 			ratingGroupDataUsages[rg] = du
-
-// 			logger.BillingLog.Error("found usage: ", flow_du.Usage, du.Usage)
-// 			logger.BillingLog.Error("found volumn: ", flow_du.TotalVol, du.TotalVol)
-// 			return
-// 		}
-// 	}
-// 	logger.BillingLog.Warnln("addVolumeToPduLevelDataUsage not found: ", flow_du)
-
-// 	filter := bson.M{"filter": "", "dnn": "", "snssai": flow_du.Snssai, "ueId": flow_du.Supi}
-// 	chargingDataInterface, err := mongoapi.RestfulAPIGetOne(chargingDataColl, filter)
-// 	if err != nil {
-// 		logger.BillingLog.Errorln(err)
-// 	}
-
-// 	var chargingRecord ChargingRecord
-// 	if json.Unmarshal(mapToByte(chargingDataInterface), &chargingRecord) != nil {
-// 		logger.BillingLog.Error(err)
-// 	}
-
-// 	if chargingRecord.RatingGroup == 0 {
-// 		logger.BillingLog.Errorf("Supi: %s's PDU Level Rating Group is 0", flow_du.Supi)
-// 		return
-// 	}
-
-// 	quota, err := strconv.Atoi(chargingRecord.Quota)
-// 	if err != nil {
-// 		logger.BillingLog.Errorf("Quota convert error, supi: %s, flow: %s", flow_du.Supi, flow_du.Filter)
-// 	}
-
-// 	ratingGroupDataUsages[chargingRecord.RatingGroup] = RatingGroupDataUsage{
-// 		Supi:      flow_du.Supi,
-// 		Filter:    "",
-// 		Snssai:    flow_du.Snssai,
-// 		Dnn:       "",
-// 		TotalVol:  flow_du.TotalVol,
-// 		DlVol:     flow_du.DlVol,
-// 		UlVol:     flow_du.UlVol,
-// 		QuotaLeft: int64(quota),
-// 		Usage:     flow_du.Usage,
-// 	}
-// }
