+++ backend/WebUI/models_charging_record.go
@@ -11,26 +11,3 @@ type ChargingRecord struct {
 	UnitCost       string `json:"unitCost,omitempty" yaml:"unitCost" bson:"unitCost" mapstructure:"unitCost"`
 	RatingGroup    int64  `json:"ratingGroup,omitempty" yaml:"ratingGroup" bson:"ratingGroup" mapstructure:"ratingGroup"`
 }
-
-// export interface FlowChargingRecord {
-//    'Supi'?: string;
-//    'Snssai'?: string;
-//    'Dnn'?: string;
-//    'Filter'?: string;
-//    'QuotaLeft'?: string;
-//    'TotalVol'?: string;
-//    'UlVol'?: string;
-//    'DlVol'?: string;
-// }
-
-// type RatingGroupDataUsage struct {
-// 	Supi      string `bson:"Supi"`
-// 	Filter    string `bson:"Filter"`
-// 	Snssai    string `bson:"Snssai"`
-// 	Dnn       string `bson:"Dnn"`
-
-// 	TotalVol  int64  `bson:"TotalVol"`
-// 	UlVol     int64  `bson:"UlVol"`
-// 	DlVol     int64  `bson:"DlVol"`
-// 	QuotaLeft int64  `bson:"QuotaLeft"`
-// }
