+++ frontend/src/axios.tsx
@@ -1,14 +1,13 @@
 import axios from "axios";
 
 const apiConfig = {
-  API_URL: '',
+  API_URL: "",
 };
 
-if (process.env.NODE_ENV === 'development') {
-  // apiConfig.API_URL = "http://127.0.0.1:5000";
-  apiConfig.API_URL = "http://10.10.0.169:5000";
+if (process.env.NODE_ENV === "development") {
+  apiConfig.API_URL = "http://127.0.0.1:5000";
 } else {
-  apiConfig.API_URL = process.env.REACT_APP_HTTP_API_URL ? process.env.REACT_APP_HTTP_API_URL : '';
+  apiConfig.API_URL = process.env.REACT_APP_HTTP_API_URL ? process.env.REACT_APP_HTTP_API_URL : "";
 }
 const instance = axios.create({
   baseURL: apiConfig.API_URL,
